tests/requirements.txt
djangorestframework==3.15.1
django-extensions==3.2.1
Django==4.1.7
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2==2.9.3
psycopg2-binary==2.9.5
bandit==1.7.5
wemake-python-styleguide==0.18.0
flake8==6.1.0
flake8-bandit==4.1.1
python-dotenv==0.21.0
django-storages==1.14.3
django-minio-backend==3.6.0
boto3==1.34.101
drf-yasg==1.21.7
pillow